blogType관련 코드 features->entities로 이동
