+++ src/entities/blog/model/type.ts
@@ -1,0 +1,47 @@
+import { CategoryType } from "entities/category";
+
+export type BlogType = PostBlogType & {
+  create_at: string;
+  blog_hashtag: BlogHashtagType[];
+  category_large: CategoryType;
+  category_middle: CategoryType;
+};
+
+export type PostBlogType = {
+  id?: string;
+  subject: string;
+  summary: string;
+  content: string;
+  thumbnail: string;
+  large_category_id: string;
+  middle_category_id: string;
+  path: string;
+};
+
+export type BlogJpType = PostBlogJpType & {
+  id: string;
+  create_at: string;
+  subject: string;
+  thumbnail: string;
+  blog_hashtag: BlogHashtagType[];
+  large_category_id: string;
+  middle_category_id: string;
+  category_large: CategoryType;
+  category_middle: CategoryType;
+};
+
+export type PostBlogJpType = {
+  path?: string;
+  subject: string;
+  summary: string;
+  content: string;
+  locale: string;
+  blog_id: string;
+};
+
+export type BlogHashtagType = {
+  hashtag_id: {
+    id: string;
+    name: string;
+  };
+};
