--- src/features/Blog/model/type.ts
@@ -1,23 +1,0 @@
-import { CategoryType } from "entities/category";
-
-export type BlogType = {
-  id?: string;
-  create_at?: string;
-  subject: string;
-  summary: string;
-  content: string;
-  thumbnail: string;
-  path: string;
-  large_category_id: string;
-  middle_category_id: string;
-  blog_hashtag?: BlogHashtagType[];
-  category_large?: CategoryType;
-  category_middle?: CategoryType;
-};
-
-export type BlogHashtagType = {
-  hashtag_id: {
-    id: string;
-    name: string;
-  };
-};
